// src/types/UIType.ts
export type ButtonSizeType =
  | 'X-LARGE'
  | 'LARGE'
  | 'MEDIUM'
  | 'SMALL'
  | 'X-SMALL';

export type ButtonType = 'primary' | 'secondary' | 'outline';

export type InputSize = 'large' | 'medium' | 'small';

// modal
export type ModalSize = 'small' | 'medium' | 'large' | 'x-large';
export type IconType = 'completed' | 'fail' | 'issue';